+++ test/fixtures/cookbooks/test/recipes/filter_lines_between.rb
@@ -6,7 +6,7 @@ directory '/tmp'
 
 # ==================== between filter =================
 template '/tmp/between' do
-  source 'dangerfile3.erb'
+  source 'samplefile3.erb'
 end
 
 filter_lines 'Change lines between matches' do