+++ test/integration/delete_lines/inspec/delete_lines.rb
@@ -1,23 +1,23 @@
 title 'Delete lines'
 
-describe file('/tmp/dangerfile1') do
+describe file('/tmp/samplefile1') do
   its(:content) { should_not match(/HI THERE I AM DANGERFILE/) }
   its(:content) { should match(/int main/) }
 end
 
-describe file('/tmp/dangerfile2') do
+describe file('/tmp/samplefile2') do
   its(:content) { should_not match(/# authorized_keys/) }
   its(:content) { should match(/ssh-rsa.*keepme@/) }
   its(:content) { should match(/ssh-dsa/) }
   its(:content) { should match(/ssh-rsa.*keepmetoo/) }
 end
 
-describe file('/tmp/dangerfile1-regexp') do
+describe file('/tmp/samplefile1-regexp') do
   its(:content) { should_not match(/HI THERE I AM DANGERFILE/) }
   its(:content) { should match(/int main/) }
 end
 
-describe file('/tmp/dangerfile2-regexp') do
+describe file('/tmp/samplefile2-regexp') do
   its(:content) { should_not match(/# authorized_keys/) }
   its(:content) { should match(/ssh-rsa.*keepme@/) }
   its(:content) { should match(/ssh-dsa/) }