+++ test/integration/filter_lines/controls/filter_lines_inline_spec.rb
@@ -8,11 +8,11 @@ control 'filter_lines - Verify the code to use adhoc filters.' do
   # ============ inline
 
   # do nothing
-  describe file('/tmp/dangerfile') do
+  describe file('/tmp/samplefile') do
     its(:content) { should match(/^HELLO THERE I AM DANGERFILE.*last line#{eol}$/m) }
   end
 
-  describe file_ext('/tmp/dangerfile') do
+  describe file_ext('/tmp/samplefile') do
     its('size_lines') { should eq 5 }
   end